requirements.txt
python-telegram-bot==21.0.1
httpx==0.27.0
aiosqlite==0.20.0
pydantic-settings==2.1.0
structlog
python-dotenv==1.0.1
pdfplumber==0.11.0
python-docx==1.1.0
Pillow==12.1.1
